# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 0trace API
  version: 0.1.0
  description: Backend API for the 0trace hacking simulation game.
paths:
...
  /api/v1/websites/:
    get:
      operationId: websites_list
      tags:
      - browser
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Website'
          description: ''
    post:
      operationId: websites_create
      tags:
      - browser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Website'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Website'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Website'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Website'
          description: ''
  /api/v1/websites/{website_id}/:
    get:
      operationId: websites_retrieve
      parameters:
      - in: path
        name: website_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - browser
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Website'
          description: ''
    patch:
      operationId: websites_partial_update
      parameters:
      - in: path
        name: website_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - browser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedWebsite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedWebsite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedWebsite'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Website'
          description: ''
    delete:
      operationId: websites_destroy
      parameters:
      - in: path
        name: website_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - browser
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    PatchedWebsite:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        title:
          type: string
          maxLength: 160
        site_type:
          allOf:
          - $ref: '#/components/schemas/SiteTypeEnum'
          readOnly: true
        trust_level:
          type: string
          readOnly: true
        html:
          type: string
        css:
          type: string
        js:
          type: string
        sandbox_policy:
          readOnly: true
        status:
          type: string
          maxLength: 20
        content_version:
          type: integer
          readOnly: true
        owner:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        domain:
          type: string
          format: uuid
          nullable: true
...
    SiteTypeEnum:
      enum:
      - predefined
      - user_hosted
      type: string
      description: |-
        * `predefined` - Predefined
        * `user_hosted` - User Hosted
...
    Website:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        title:
          type: string
          maxLength: 160
        site_type:
          allOf:
          - $ref: '#/components/schemas/SiteTypeEnum'
          readOnly: true
        trust_level:
          type: string
          readOnly: true
        html:
          type: string
        css:
          type: string
        js:
          type: string
        sandbox_policy:
          readOnly: true
        status:
          type: string
          maxLength: 20
        content_version:
          type: integer
          readOnly: true
        owner:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        domain:
          type: string
          format: uuid
          nullable: true
      required:
      - content_version
      - created_at
      - id
      - owner
      - sandbox_policy
      - site_type
      - title
      - trust_level
      - updated_at
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...